requirements.txt
Django>=4.2,<5.0
pymysql>=1.1.0
Pillow>=10.0.0
python-decouple>=3.8
django-cors-headers>=4.3.0
djangorestframework>=3.14.0
